Cleaned up file, moved SmartRateTest to ShipmentTest. Added root element to request so it ignores result root.
