--- EasyPostTest/SmartRateData.txt
@@ -1,1 +1,0 @@
-﻿{"result":[{"carrier":"USPS","carrier_account_id":"ca_7642d249fdcf47bcb5da9ea34c96dfcf","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":null,"delivery_date_guaranteed":false,"delivery_days":null,"est_delivery_days":null,"id":"rate_d1ecce3fae964bf3a992fbddaba4f622","list_currency":"USD","list_rate":42.75,"mode":"test","object":"Rate","rate":42.75,"retail_currency":"USD","retail_rate":49.45,"service":"Express","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":1,"percentile_75":2,"percentile_85":2,"percentile_90":3,"percentile_95":3,"percentile_97":4,"percentile_99":6},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"USPS","carrier_account_id":"ca_7642d249fdcf47bcb5da9ea34c96dfcf","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":null,"delivery_date_guaranteed":false,"delivery_days":5,"est_delivery_days":5,"id":"rate_21276e0334364beb8fd836a58467cc00","list_currency":"USD","list_rate":9.28,"mode":"test","object":"Rate","rate":9.28,"retail_currency":"USD","retail_rate":9.28,"service":"ParcelSelect","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":1,"percentile_75":2,"percentile_85":3,"percentile_90":5,"percentile_95":8,"percentile_97":9,"percentile_99":13},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"USPS","carrier_account_id":"ca_7642d249fdcf47bcb5da9ea34c96dfcf","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":null,"delivery_date_guaranteed":false,"delivery_days":2,"est_delivery_days":2,"id":"rate_24287ede57d1438c918f123e62e745df","list_currency":"USD","list_rate":9.53,"mode":"test","object":"Rate","rate":9.53,"retail_currency":"USD","retail_rate":12.5,"service":"Priority","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":2,"percentile_75":2,"percentile_85":2,"percentile_90":2,"percentile_95":3,"percentile_97":3,"percentile_99":5},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-04T23:00:00Z","delivery_date_guaranteed":false,"delivery_days":3,"est_delivery_days":3,"id":"rate_ffa6a715cb6f40c29ff5611e187fdb08","list_currency":"USD","list_rate":27.04,"mode":"test","object":"Rate","rate":25.84,"retail_currency":"USD","retail_rate":25.84,"service":"3DaySelect","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":2,"percentile_75":2,"percentile_85":2,"percentile_90":2,"percentile_95":3,"percentile_97":3,"percentile_99":3},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-02T09:30:00Z","delivery_date_guaranteed":true,"delivery_days":1,"est_delivery_days":1,"id":"rate_ba6972c362c949a98623387e64c7f8a8","list_currency":"USD","list_rate":138.93,"mode":"test","object":"Rate","rate":133.64,"retail_currency":"USD","retail_rate":133.64,"service":"NextDayAirEarlyAM","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":null,"percentile_75":null,"percentile_85":null,"percentile_90":null,"percentile_95":null,"percentile_97":null,"percentile_99":null},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-03T23:00:00Z","delivery_date_guaranteed":false,"delivery_days":2,"est_delivery_days":2,"id":"rate_f4f6e1fbd2cb4ff9bccd7d8f7b48954f","list_currency":"USD","list_rate":40.68,"mode":"test","object":"Rate","rate":37.63,"retail_currency":"USD","retail_rate":37.63,"service":"2ndDayAirAM","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":null,"percentile_75":null,"percentile_85":null,"percentile_90":null,"percentile_95":null,"percentile_97":null,"percentile_99":null},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-03T23:00:00Z","delivery_date_guaranteed":false,"delivery_days":2,"est_delivery_days":2,"id":"rate_2596f17228614fb6950ef1834ac1173f","list_currency":"USD","list_rate":14.01,"mode":"test","object":"Rate","rate":14.35,"retail_currency":"USD","retail_rate":14.35,"service":"Ground","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":2,"percentile_75":2,"percentile_85":2,"percentile_90":2,"percentile_95":3,"percentile_97":3,"percentile_99":3},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-02T23:00:00Z","delivery_date_guaranteed":false,"delivery_days":1,"est_delivery_days":1,"id":"rate_972bd0bee1474a9e8aa1f0ba9466e53e","list_currency":"USD","list_rate":104.4,"mode":"test","object":"Rate","rate":98.46,"retail_currency":"USD","retail_rate":98.46,"service":"NextDayAirSaver","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":1,"percentile_75":1,"percentile_85":1,"percentile_90":1,"percentile_95":1,"percentile_97":1,"percentile_99":2},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-02T12:00:00Z","delivery_date_guaranteed":false,"delivery_days":1,"est_delivery_days":1,"id":"rate_1acc6b58d54c4be5ae99e3d341ed4507","list_currency":"USD","list_rate":106.83,"mode":"test","object":"Rate","rate":101.54,"retail_currency":"USD","retail_rate":101.54,"service":"NextDayAir","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":1,"percentile_75":1,"percentile_85":1,"percentile_90":1,"percentile_95":1,"percentile_97":1,"percentile_99":2},"updated_at":"2021-05-31T05:37:46Z"},{"carrier":"UPS","carrier_account_id":"ca_a8f9fb8d91a840f6a2f15eb96c348ca8","created_at":"2021-05-31T05:37:46Z","currency":"USD","delivery_date":"2021-06-03T23:00:00Z","delivery_date_guaranteed":false,"delivery_days":2,"est_delivery_days":2,"id":"rate_b1243197bb76477fb702e1b02fab224d","list_currency":"USD","list_rate":36.71,"mode":"test","object":"Rate","rate":33.6,"retail_currency":"USD","retail_rate":33.6,"service":"2ndDayAir","shipment_id":"shp_af184bc8c25f4879947af2ceb36b8b98","time_in_transit":{"percentile_50":2,"percentile_75":2,"percentile_85":2,"percentile_90":2,"percentile_95":2,"percentile_97":2,"percentile_99":3},"updated_at":"2021-05-31T05:37:46Z"}]}
